bancoV2/requirements.txt
django>=5.1.6
psycopg2-binary
django-cors-headers
djangorestframework 
python-dotenv 
djangorestframework-simplejwt 